Use a file-of-filenames for mergeSTR input file arguments
Since TRTools v6.0, mergeSTR has a --vcfs-list option that takes
a list file containing filenames as an alternative to listing them
on the command line. Use this to entirely avoid line length limits
when merging thousands of VCFs.

Use cpg_utils.hail_batch.command() so that indentation is removed
from the shell snippet, in particular that on the first batch_vcfs
line and the "EOF" line (which otherwise wouldn't work). This also
sets pipefail and other error-handling options, which is helpful.

Use a newline helper variable, as Python 3.10 f-strings can't
contain backslashes such as `f"""...{"\n"...}..."""`.