Use a file-of-filenames for mergeSTR input file arguments
Since TRTools v6.0, mergeSTR has a --vcfs-list option that takes
a list file containing filenames as an alternative to listing them
on the command line. Use this to entirely avoid line length limits
when merging thousands of VCFs.

Write the `cat ... EOF` command all on one source line, so that
there is no indentation on the input data and "EOF" line in the
shell snippet, which otherwise wouldn't work. (Don't quote <<'EOF'
as the filenames usually contain ${BATCH_TMPDIR} which will need
to be expanded.)

Use a newline helper variable, as Python 3.10 f-strings can't
contain backslashes such as `f"""...{"\n"...}..."""`.